Post Merge branch 'master' corrections
