+++ src/model.ts
@@ -3,11 +3,10 @@ import { Dialog, showErrorMessage } from '@jupyterlab/apputils';
 import { IChangedArgs, PathExt } from '@jupyterlab/coreutils';
 import { ServerConnection } from '@jupyterlab/services';
 import { ISettingRegistry } from '@jupyterlab/settingregistry';
-import { Dialog, showErrorMessage } from '@jupyterlab/apputils';
+import { LinkedList } from '@lumino/collections';
 import { CommandRegistry } from '@lumino/commands';
 import { JSONObject } from '@lumino/coreutils';
 import { Poll } from '@lumino/polling';
-import { LinkedList } from '@lumino/collections';
 import { ISignal, Signal } from '@lumino/signaling';
 import { httpGitRequest } from './git';
 import { Git, IGitExtension } from './tokens';
