+++ tests/test-components/Toolbar.spec.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { ActionButton } from '../../src/components/ActionButton';
 import { Toolbar } from '../../src/components/Toolbar';
 import * as git from '../../src/git';
-import { CommandIDs } from '../../src/gitMenuCommands';
+import { CommandIDs } from '../../src/commandsAndMenu';
 import { GitExtension } from '../../src/model';
 import { pullIcon, pushIcon } from '../../src/style/icons';
 import {
